+++ demos/chinook/src/main/java/is/codion/framework/demos/chinook/ui/AlbumTablePanel.java
@@ -50,7 +50,7 @@ public final class AlbumTablePanel extends EntityTablePanel {
 		super(tableModel, config -> config
 						.editComponentFactory(Album.TAGS, new TagEditComponentFactory())
 						.table(builder -> builder.cellRenderer(Album.RATING,
-										() -> EntityTableCellRenderer.builder(Album.RATING, tableModel)
+										EntityTableCellRenderer.builder(Album.RATING, tableModel)
 														.string(RATINGS::get)
 														.toolTipData(true)
 														.build())));