+++ changelog.md
@@ -44,6 +44,7 @@ Codion Change Log
 - FilterTable.Builder.cellEditor() and cellRenderer() added, related refactoring.
 - FilterTableCellRenderer.Color.Provider.color() column parameter added.
 - FilterTableCellRenderer.Builder row and column identifier type parameters added.
+- FilterTable.Builder.cellEditor() and cellRenderer() now long Supplier based.
 ### is.codion.framework.domain
 - Entity.copy(), copyBuilder() and deepCopy() moved to Entity.Copy, renamed mutable(), builder() and deep() respectively, immutable() added.
 - ImmutableEntity bug fixed, StackOverflowError when creating an immutable entity with cyclical foreign key dependencies, some refactoring.